requirements.txt
customtkinter>=5.0.0
Pillow>=9.0.0
